feat: Update w/ menu changes, preventing layout on menu open, etc.
Bring live site up to date with latest commit of personal-site repository. This should include changes to the menu such that no significant layout changes are required to render the menu opening, only some minor layout and compositing for the menu itself. Also, change the menu entries to update the jumbotron on click rather than hover